# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/bookings/all": {
            "get": {
                "description": "Retrieves all bookings, with the option to filter by pricing mode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Get all bookings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Capability to filter by pricing mode",
                        "name": "Capability",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - Return all bookings in pricing mode",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Booking"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bookings/confirm/{id}": {
            "put": {
                "description": "Confirms a booking by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Confirm a booking",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID to confirm",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking confirmed successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Booking not found after confirmation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bookings/post": {
            "post": {
                "description": "Creates a new booking and updates the availability accordingly",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Post a booking",
                "parameters": [
                    {
                        "description": "Request Payload for Posting a Booking",
                        "name": "BookingPayload_Rq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BookingPayload_Rq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Booking successfully created",
                        "schema": {
                            "$ref": "#/definitions/model.Booking"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bookings/{id}": {
            "get": {
                "description": "Fetches a booking by its ID, with the option to filter by pricing mode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Get a booking by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Capability to filter by pricing mode",
                        "name": "Capability",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - Return booking in pricing mode",
                        "schema": {
                            "$ref": "#/definitions/model.Booking"
                        }
                    },
                    "404": {
                        "description": "Booking not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Booking": {
            "type": "object",
            "properties": {
                "availabilityId": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "units": {
                    "type": "integer"
                }
            }
        },
        "model.BookingPayload_Rq": {
            "type": "object",
            "properties": {
                "availabilityId": {
                    "type": "string"
                },
                "productId": {
                    "type": "string"
                },
                "units": {
                    "type": "integer"
                }
            }
        },
...
    }
}